Reuse existing SlotAnnotation objects for bulk loads

+++ src/main/java/org/alliancegenome/curation_api/services/validation/dto/AlleleDTOValidator.java
@@ -1,7 +1,11 @@
 package org.alliancegenome.curation_api.services.validation.dto;
 
 import java.util.ArrayList;
+import java.util.HashMap;
 import java.util.List;
+import java.util.Map;
+import java.util.function.Function;
+import java.util.stream.Collectors;
 
 import javax.enterprise.context.RequestScoped;
 import javax.inject.Inject;
@@ -35,6 +39,7 @@ import org.alliancegenome.curation_api.model.ingest.dto.AlleleSecondaryIdSlotAnn
 import org.alliancegenome.curation_api.model.ingest.dto.NameSlotAnnotationDTO;
 import org.alliancegenome.curation_api.response.ObjectResponse;
 import org.alliancegenome.curation_api.services.ReferenceService;
+import org.alliancegenome.curation_api.services.helpers.slotAnnotations.SlotAnnotationIdentityHelper;
 import org.alliancegenome.curation_api.services.validation.dto.base.BaseDTOValidator;
 import org.alliancegenome.curation_api.services.validation.dto.slotAnnotations.alleleSlotAnnotations.AlleleFullNameSlotAnnotationDTOValidator;
 import org.alliancegenome.curation_api.services.validation.dto.slotAnnotations.alleleSlotAnnotations.AlleleInheritanceModeSlotAnnotationDTOValidator;
@@ -80,6 +85,8 @@ public class AlleleDTOValidator extends BaseDTOValidator {
 	AlleleSynonymSlotAnnotationDTOValidator alleleSynonymDtoValidator;
 	@Inject
 	AlleleSecondaryIdSlotAnnotationDTOValidator alleleSecondaryIdDtoValidator;
+	@Inject
+	SlotAnnotationIdentityHelper slotAnnotationIdentity;
 
 	@Transactional
 	public Allele validateAlleleDTO(AlleleDTO dto) throws ObjectValidationException {
@@ -129,57 +136,73 @@ public class AlleleDTOValidator extends BaseDTOValidator {
 			allele.setReferences(null);
 		}
 
+		Map<String, AlleleMutationTypeSlotAnnotation> existingMutationTypes = new HashMap<>();
 		if (CollectionUtils.isNotEmpty(allele.getAlleleMutationTypes())) {
-			allele.getAlleleMutationTypes().forEach(amt -> {
-				amt.setSingleAllele(null);
-				alleleMutationTypeDAO.remove(amt.getId());
-			});
+			for (AlleleMutationTypeSlotAnnotation existingMutationType : allele.getAlleleMutationTypes()) {
+				existingMutationTypes.put(slotAnnotationIdentity.alleleMutationTypesIdentity(existingMutationType), existingMutationType);
+			}
 		}
-
+		
 		List<AlleleMutationTypeSlotAnnotation> mutationTypes = new ArrayList<>();
+		List<String> mutationTypeIdentities = new ArrayList<>();
 		if (CollectionUtils.isNotEmpty(dto.getAlleleMutationTypeDtos())) {
 			for (AlleleMutationTypeSlotAnnotationDTO mutationTypeDTO : dto.getAlleleMutationTypeDtos()) {
-				ObjectResponse<AlleleMutationTypeSlotAnnotation> mutationTypeResponse = alleleMutationTypeDtoValidator.validateAlleleMutationTypeSlotAnnotationDTO(mutationTypeDTO);
+				ObjectResponse<AlleleMutationTypeSlotAnnotation> mutationTypeResponse = alleleMutationTypeDtoValidator.validateAlleleMutationTypeSlotAnnotationDTO(existingMutationTypes.get(slotAnnotationIdentity.alleleMutationTypesDtoIdentity(mutationTypeDTO)), mutationTypeDTO);
 				if (mutationTypeResponse.hasErrors()) {
 					alleleResponse.addErrorMessage("allele_mutation_type_dtos", mutationTypeResponse.errorMessagesString());
 				} else {
 					AlleleMutationTypeSlotAnnotation mutationType = mutationTypeResponse.getEntity();
 					mutationTypes.add(mutationType);
+					mutationTypeIdentities.add(slotAnnotationIdentity.alleleMutationTypesIdentity(mutationType));
 				}
 			}
 		}
 		
-		if (CollectionUtils.isNotEmpty(allele.getAlleleInheritanceModes())) {
-			allele.getAlleleInheritanceModes().forEach(aim -> {
-				aim.setSingleAllele(null);
-				alleleInheritanceModeDAO.remove(aim.getId());
+		if (!existingMutationTypes.isEmpty()) {
+			existingMutationTypes.forEach((k,v) -> {
+				if (!mutationTypeIdentities.contains(k)) {
+					v.setSingleAllele(null);
+					alleleMutationTypeDAO.remove(v.getId());
+				}
 			});
 		}
-
+		
+		Map<String, AlleleInheritanceModeSlotAnnotation> existingInheritanceModes = new HashMap<>();
+		if (CollectionUtils.isNotEmpty(allele.getAlleleInheritanceModes())) {
+			for (AlleleInheritanceModeSlotAnnotation existingInheritanceMode : allele.getAlleleInheritanceModes()) {
+				existingInheritanceModes.put(slotAnnotationIdentity.alleleInheritanceModeIdentity(existingInheritanceMode), existingInheritanceMode);
+			}
+		}
+		
 		List<AlleleInheritanceModeSlotAnnotation> inheritanceModes = new ArrayList<>();
+		List<String> inheritanceModeIdentities = new ArrayList<>();
 		if (CollectionUtils.isNotEmpty(dto.getAlleleInheritanceModeDtos())) {
 			for (AlleleInheritanceModeSlotAnnotationDTO inheritanceModeDTO : dto.getAlleleInheritanceModeDtos()) {
-				ObjectResponse<AlleleInheritanceModeSlotAnnotation> inheritanceModeResponse = alleleInheritanceModeDtoValidator.validateAlleleInheritanceModeSlotAnnotationDTO(inheritanceModeDTO);
+				ObjectResponse<AlleleInheritanceModeSlotAnnotation> inheritanceModeResponse = alleleInheritanceModeDtoValidator.validateAlleleInheritanceModeSlotAnnotationDTO(existingInheritanceModes.get(slotAnnotationIdentity.alleleInheritanceModeDtoIdentity(inheritanceModeDTO)), inheritanceModeDTO);
 				if (inheritanceModeResponse.hasErrors()) {
 					alleleResponse.addErrorMessage("allele_inheritance_mode_dtos", inheritanceModeResponse.errorMessagesString());
 				} else {
 					AlleleInheritanceModeSlotAnnotation inheritanceMode = inheritanceModeResponse.getEntity();
 					inheritanceModes.add(inheritanceMode);
+					inheritanceModeIdentities.add(slotAnnotationIdentity.alleleInheritanceModeIdentity(inheritanceMode));
 				}
 			}
 		}
-
-		if (allele.getAlleleSymbol() != null) {
-			AlleleSymbolSlotAnnotation symbol = allele.getAlleleSymbol();
-			symbol.setSingleAllele(null);
-			alleleSymbolDAO.remove(symbol.getId());
+		
+		if (!existingInheritanceModes.isEmpty()) {
+			existingInheritanceModes.forEach((k,v) -> {
+				if (!inheritanceModeIdentities.contains(k)) {
+					v.setSingleAllele(null);
+					alleleInheritanceModeDAO.remove(v.getId());
+				}
+			});
 		}
 
-		AlleleSymbolSlotAnnotation symbol = null;
+		AlleleSymbolSlotAnnotation symbol = allele.getAlleleSymbol();
 		if (dto.getAlleleSymbolDto() == null) {
 			alleleResponse.addErrorMessage("allele_symbol_dto", ValidationConstants.REQUIRED_MESSAGE);
 		} else {
-			ObjectResponse<AlleleSymbolSlotAnnotation> symbolResponse = alleleSymbolDtoValidator.validateAlleleSymbolSlotAnnotationDTO(dto.getAlleleSymbolDto());
+			ObjectResponse<AlleleSymbolSlotAnnotation> symbolResponse = alleleSymbolDtoValidator.validateAlleleSymbolSlotAnnotationDTO(symbol, dto.getAlleleSymbolDto());
 			if (symbolResponse.hasErrors()) {
 				alleleResponse.addErrorMessage("allele_symbol_dto", symbolResponse.errorMessagesString());
 			} else {
@@ -187,33 +210,31 @@ public class AlleleDTOValidator extends BaseDTOValidator {
 			}
 		}
 
-		if (allele.getAlleleFullName() != null) {
-			AlleleFullNameSlotAnnotation fullName = allele.getAlleleFullName();
+		AlleleFullNameSlotAnnotation fullName = allele.getAlleleFullName();
+		if (allele.getAlleleFullName() != null && dto.getAlleleFullNameDto() == null) {
 			fullName.setSingleAllele(null);
 			alleleFullNameDAO.remove(fullName.getId());
 		}
 
-		AlleleFullNameSlotAnnotation fullName = null;
 		if (dto.getAlleleFullNameDto() != null) {
-			ObjectResponse<AlleleFullNameSlotAnnotation> fullNameResponse = alleleFullNameDtoValidator.validateAlleleFullNameSlotAnnotationDTO(dto.getAlleleFullNameDto());
+			ObjectResponse<AlleleFullNameSlotAnnotation> fullNameResponse = alleleFullNameDtoValidator.validateAlleleFullNameSlotAnnotationDTO(fullName, dto.getAlleleFullNameDto());
 			if (fullNameResponse.hasErrors()) {
 				alleleResponse.addErrorMessage("allele_full_name_dto", fullNameResponse.errorMessagesString());
 			} else {
 				fullName = fullNameResponse.getEntity();
 			}
+		} else {
+			fullName = null;
 		}
 
-		if (CollectionUtils.isNotEmpty(allele.getAlleleSynonyms())) {
-			allele.getAlleleSynonyms().forEach(as -> {
-				as.setSingleAllele(null);
-				alleleSynonymDAO.remove(as.getId());
-			});
-		}
-
+		Map<String, AlleleSynonymSlotAnnotation> existingSynonyms = new HashMap<>();
+		if (CollectionUtils.isNotEmpty(allele.getAlleleSynonyms()))
+			existingSynonyms = allele.getAlleleSynonyms().stream().collect(Collectors.toMap(AlleleSynonymSlotAnnotation::getFormatText, Function.identity()));
+		
 		List<AlleleSynonymSlotAnnotation> synonyms = new ArrayList<>();
 		if (CollectionUtils.isNotEmpty(dto.getAlleleSynonymDtos())) {
 			for (NameSlotAnnotationDTO synonymDTO : dto.getAlleleSynonymDtos()) {
-				ObjectResponse<AlleleSynonymSlotAnnotation> synonymResponse = alleleSynonymDtoValidator.validateAlleleSynonymSlotAnnotationDTO(synonymDTO);
+				ObjectResponse<AlleleSynonymSlotAnnotation> synonymResponse = alleleSynonymDtoValidator.validateAlleleSynonymSlotAnnotationDTO(existingSynonyms.get(synonymDTO.getFormatText()), synonymDTO);
 				if (synonymResponse.hasErrors()) {
 					alleleResponse.addErrorMessage("allele_synonym_dtos", synonymResponse.errorMessagesString());
 				} else {
@@ -222,18 +243,25 @@ public class AlleleDTOValidator extends BaseDTOValidator {
 				}
 			}
 		}
-
-		if (CollectionUtils.isNotEmpty(allele.getAlleleSecondaryIds())) {
-			allele.getAlleleSecondaryIds().forEach(asid -> {
-				asid.setSingleAllele(null);
-				alleleSecondaryIdDAO.remove(asid.getId());
+		
+		if (!existingSynonyms.isEmpty()) {
+			List<String> synonymNames = synonyms.stream().map(AlleleSynonymSlotAnnotation::getFormatText).collect(Collectors.toList());
+			existingSynonyms.forEach((k,v) -> {
+				if (!synonymNames.contains(k)) {
+					v.setSingleAllele(null);
+					alleleSynonymDAO.remove(v.getId());
+				}
 			});
 		}
 
+		Map<String, AlleleSecondaryIdSlotAnnotation> existingSecondaryIds = new HashMap<>();
+		if (CollectionUtils.isNotEmpty(allele.getAlleleSecondaryIds()))
+			existingSecondaryIds = allele.getAlleleSecondaryIds().stream().collect(Collectors.toMap(AlleleSecondaryIdSlotAnnotation::getSecondaryId, Function.identity()));
+		
 		List<AlleleSecondaryIdSlotAnnotation> secondaryIds = new ArrayList<>();
 		if (CollectionUtils.isNotEmpty(dto.getAlleleSecondaryIdDtos())) {
 			for (AlleleSecondaryIdSlotAnnotationDTO secondaryIdDTO : dto.getAlleleSecondaryIdDtos()) {
-				ObjectResponse<AlleleSecondaryIdSlotAnnotation> secondaryIdResponse = alleleSecondaryIdDtoValidator.validateAlleleSecondaryIdSlotAnnotationDTO(secondaryIdDTO);
+				ObjectResponse<AlleleSecondaryIdSlotAnnotation> secondaryIdResponse = alleleSecondaryIdDtoValidator.validateAlleleSecondaryIdSlotAnnotationDTO(existingSecondaryIds.get(secondaryIdDTO.getSecondaryId()), secondaryIdDTO);
 				if (secondaryIdResponse.hasErrors()) {
 					alleleResponse.addErrorMessage("allele_secondary_id_dtos", secondaryIdResponse.errorMessagesString());
 				} else {
@@ -241,6 +269,16 @@ public class AlleleDTOValidator extends BaseDTOValidator {
 					secondaryIds.add(secondaryId);
 				}
 			}
+		}
+		
+		if (!existingSecondaryIds.isEmpty()) {
+			List<String> secondaryIdNames = secondaryIds.stream().map(AlleleSecondaryIdSlotAnnotation::getSecondaryId).collect(Collectors.toList());
+			existingSecondaryIds.forEach((k,v) -> {
+				if (!secondaryIdNames.contains(k)) {
+					v.setSingleAllele(null);
+					alleleSecondaryIdDAO.remove(v.getId());
+				}
+			});
 		}
 
 		if (alleleResponse.hasErrors()) {
